requirements.txt
streamlit
groq
fredapi
pandas
plotly
python-dotenv